minio_storage/minio_requirements.txt
minio==7.2.5
python-dotenv==1.0.1